:recycle: REFACTOR :Modify GardenCategories Column attribute

+++ wantPlant/src/main/java/umc/wantPlant/garden/domain/dto/GardenRequestDTO.java
@@ -14,8 +14,8 @@ public class GardenRequestDTO {
 
 		@NotBlank
 		private String description;
-		
+
 		@NotNull
-		private Long gardenCategory;
+		private Integer gardenCategory;
 	}
 }
